+++ src/main/java/org/boredombuster/chessengine/Board.java
@@ -1,0 +1,5 @@
+package org.boredombuster.chessengine;
+
+public class Board {
+
+}
