+++ src/main/java/org/boredombuster/controller/TriviaController.java
@@ -1,13 +1,12 @@
-package org.informationblitz.controller;
+package org.boredombuster.controller;
 
-import org.informationblitz.dto.TriviaDTO;
-import org.informationblitz.service.TriviaService;
+import org.boredombuster.dto.TriviaDTO;
+import org.boredombuster.service.TriviaService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
