+++ src/main/java/org/boredombuster/dto/TriviaDTO.java
@@ -1,4 +1,4 @@
-package org.informationblitz.dto;
+package org.boredombuster.dto;
 
 public class TriviaDTO {
     private String category;
