+++ src/main/java/org/boredombuster/service/WeatherService.java
@@ -1,10 +1,7 @@
-package org.informationblitz.service;
+package org.boredombuster.service;
 
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.informationblitz.controller.WeatherController;
-import org.informationblitz.dto.WeatherDTO;
+import org.boredombuster.dto.WeatherDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -17,7 +14,6 @@ import java.net.HttpURLConnection;
 import java.net.URL;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
-import java.util.Scanner;
 import java.util.stream.Collectors;
 
 @Service
